set up typescript
