+++ nodemon.json
@@ -1,0 +1,6 @@
+{
+	"watch": ["src"],
+	"ext": ".ts,.js",
+	"ignore": [],
+	"exec": "npx ts-node ./src/index.ts"
+}
