+++ package.json
@@ -17,6 +17,10 @@
 		"swagger-ui-express": "^5.0.0"
 	},
 	"devDependencies": {
-		"nodemon": "^3.0.1"
+		"@types/express": "^4.17.18",
+		"@types/node": "^20.8.3",
+		"@types/swagger-ui-express": "^4.1.4",
+		"nodemon": "^3.0.1",
+		"typescript": "^5.2.2"
 	}
 }
