--- src/index.js
@@ -1,13 +1,0 @@
-const express = require('express');
-const app = express();
-const PORT = 3000;
-const swaggerUi = require('swagger-ui-express');
-const swaggerJsdoc = require('./swaggerConfig');
-const routes = require('./routes/routes');
-
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerJsdoc));
-app.use(routes);
-
-app.listen(PORT, () => {
-	console.log(`Server is listening on port ${PORT}`);
-});
