+++ src/swaggerConfig.js
@@ -16,7 +16,7 @@ const options = {
 			},
 		],
 	},
-	apis: ['./src/routes/*.js'],
+	apis: ['./src/routes/*.ts'],
 };
 
 const swaggerSpec = swaggerJsdoc(options);
