+++ tsconfig.json
@@ -1,0 +1,16 @@
+{
+	"compilerOptions": {
+		"target": "es6",
+		"module": "commonjs",
+		"lib": ["es6"],
+		"moduleResolution": "node",
+		"allowJs": true,
+		"outDir": "./build",
+		"sourceMap": true,
+		"strict": true,
+		"noImplicitAny": true,
+		"esModuleInterop": true,
+		"resolveJsonModule": true,
+	},
+	"exclude": ["./node_modules"]
+}
